backend/requirements.txt
Flask>=2.2
flask-socketio>=5.3.0
python-dotenv>=1.0
openai>=0.27.0
flask-jwt-extended>=4.4.4
pymongo>=4.3.3
bcrypt>=4.0.1
flask-cors>=3.0.10
python-socketio>=5.7.2
